+++ packages/lib/src/utils/table.rs
@@ -2,22 +2,22 @@ use std::future::Future;
 
 use mlua::prelude::*;
 
-pub struct TableBuilder<'lua> {
-    lua: &'lua Lua,
-    tab: LuaTable<'lua>,
+pub struct TableBuilder {
+    lua: &'static Lua,
+    tab: LuaTable<'static>,
 }
 
 #[allow(dead_code)]
-impl<'lua> TableBuilder<'lua> {
-    pub fn new(lua: &'lua Lua) -> LuaResult<Self> {
+impl TableBuilder {
+    pub fn new(lua: &'static Lua) -> LuaResult<Self> {
         let tab = lua.create_table()?;
         Ok(Self { lua, tab })
     }
 
     pub fn with_value<K, V>(self, key: K, value: V) -> LuaResult<Self>
     where
-        K: ToLua<'lua>,
-        V: ToLua<'lua>,
+        K: ToLua<'static>,
+        V: ToLua<'static>,
     {
         self.tab.raw_set(key, value)?;
         Ok(self)
@@ -25,8 +25,8 @@ impl<'lua> TableBuilder<'lua> {
 
     pub fn with_values<K, V>(self, values: Vec<(K, V)>) -> LuaResult<Self>
     where
-        K: ToLua<'lua>,
-        V: ToLua<'lua>,
+        K: ToLua<'static>,
+        V: ToLua<'static>,
     {
         for (key, value) in values {
             self.tab.raw_set(key, value)?;
@@ -36,7 +36,7 @@ impl<'lua> TableBuilder<'lua> {
 
     pub fn with_sequential_value<V>(self, value: V) -> LuaResult<Self>
     where
-        V: ToLua<'lua>,
+        V: ToLua<'static>,
     {
         self.tab.raw_push(value)?;
         Ok(self)
@@ -44,7 +44,7 @@ impl<'lua> TableBuilder<'lua> {
 
     pub fn with_sequential_values<V>(self, values: Vec<V>) -> LuaResult<Self>
     where
-        V: ToLua<'lua>,
+        V: ToLua<'static>,
     {
         for value in values {
             self.tab.raw_push(value)?;
@@ -59,10 +59,10 @@ impl<'lua> TableBuilder<'lua> {
 
     pub fn with_function<K, A, R, F>(self, key: K, func: F) -> LuaResult<Self>
     where
-        K: ToLua<'lua>,
-        A: FromLuaMulti<'lua>,
-        R: ToLuaMulti<'lua>,
-        F: 'static + Fn(&'lua Lua, A) -> LuaResult<R>,
+        K: ToLua<'static>,
+        A: FromLuaMulti<'static>,
+        R: ToLuaMulti<'static>,
+        F: 'static + Fn(&'static Lua, A) -> LuaResult<R>,
     {
         let f = self.lua.create_function(func)?;
         self.with_value(key, LuaValue::Function(f))
@@ -70,22 +70,22 @@ impl<'lua> TableBuilder<'lua> {
 
     pub fn with_async_function<K, A, R, F, FR>(self, key: K, func: F) -> LuaResult<Self>
     where
-        K: ToLua<'lua>,
-        A: FromLuaMulti<'lua>,
-        R: ToLuaMulti<'lua>,
-        F: 'static + Fn(&'lua Lua, A) -> FR,
-        FR: 'lua + Future<Output = LuaResult<R>>,
+        K: ToLua<'static>,
+        A: FromLuaMulti<'static>,
+        R: ToLuaMulti<'static>,
+        F: 'static + Fn(&'static Lua, A) -> FR,
+        FR: 'static + Future<Output = LuaResult<R>>,
     {
         let f = self.lua.create_async_function(func)?;
         self.with_value(key, LuaValue::Function(f))
     }
 
-    pub fn build_readonly(self) -> LuaResult<LuaTable<'lua>> {
+    pub fn build_readonly(self) -> LuaResult<LuaTable<'static>> {
         self.tab.set_readonly(true);
         Ok(self.tab)
     }
 
-    pub fn build(self) -> LuaResult<LuaTable<'lua>> {
+    pub fn build(self) -> LuaResult<LuaTable<'static>> {
         Ok(self.tab)
     }
 }